GridObserver created to transmit grid state, simulation object create

+++ backend/LawsLaboratory/LawsLaboratory.Application/Simulation/Configuration/RuntimeConfiguration/TimeConfiguration.cs
@@ -17,7 +17,7 @@
 namespace LawsLaboratory.Application.Simulation.Configuration.RuntimeConfigurationPack;
     public sealed record TimeConfiguration
     {
-        public int? MaxCycles { get; init; }
+        public long? MaxCycles { get; init; }
 
         public int? DelayMsPerCycle { get; init; }
     }
